Update to 0.1.14
This update fixes a bug in get_molecular_weight() and corrects for numerical precision issues in get_isoelectric_point(). It also updates all the pH dependent functions to include Tyr and Cys.

In addition, new tests were written to test these new functions.

Also, the versioning has be removed from each file, basically because it's a huge pain in the ass to update this on EACH update and I don't trust SED scripts to not bork everything in case the regex gets included somewhere in a file...

0.1.14 will be pushed to PyPI upon this commit.