Removed 'use_2to3' and main_test
Updated setup.py to remove use_2to3 from the extras, as it is no
longer needed, and it caused setup to fail under Python 3.6.

Also removed localcider/tests/main_test.py, because it is nowhere
referenced directly, and after treatment with 2to3, it was identical
with localcider/tests/__init__.py.